+++ RELEASE_NOTES.md
@@ -9,6 +9,7 @@
 **Improvements**
 * Store Player information from instance list directly into SQlite to improve speed when
 opening a user profile.
+* The "Taskbar" is no longer part of the Modal overlay layer so you can move the window even if a modal is open now
 
 **Fixes**
 * Fixed the Group Modal not showing the group creation date.